requirements-dev.txt
-r requirements.txt
WebOb==1.2.3
WebTest==2.0.3
Werkzeug==0.8.3
beautifulsoup4==4.1.3
coverage==3.6
django-debug-toolbar==0.9.4
django-discover-runner==0.3
django-webtest==1.6
factory-boy==1.2.0
flake8==2.0
mccabe==0.2
mock==1.0.1
ordereddict==1.1
pep8==1.4.5
pep8-naming==0.2.1
pyflakes==0.6.1
waitress==0.8.2
